Add files via upload
Updated stuff. Don't really remember what. Think I made more variables for the future to use. Also I think I added velocity junk to the monster

+++ src/main/java/com/johnnyong/android/gamedevbumperhero/ActionItem.kt
@@ -1,5 +1,5 @@
 package com.johnnyong.android.gamedevbumperhero
 
 interface ActionItem {
-    fun doClick(px:Int, py:Int): Boolean
+    fun doClick(px: Double, py:Int): Boolean
 }
